refactor(examples): consolidate and organize example scripts (#148)
Reorganize examples into logical directories:
- benchmarks/: Performance benchmarks (matmul, CUDA Graph)
- chat/: Chat CLI applications (standard, MoE, thinking, Triton)
- demos/archived/: Version-specific demos (v01-v026) for reference

Keep current demos at top level:
- demo_gpu.py, demo_cuda_graph.py, demo_llm_e2e.py, etc.

Update README.md with new structure and usage instructions.
