+++ examples/benchmarks/benchmark_compare.py
