+++ examples/chat/chat_cli_thinking.py
